+++ src/main/kotlin/extensions/Provider.kt
@@ -1,0 +1,5 @@
+package ru.endlesscode.bukkitgradle.extensions
+
+import org.gradle.api.provider.Property
+
+internal fun <T> Property<T>.finalizedOnRead(): Property<T> = apply { finalizeValueOnRead() }