+++ src/main/kotlin/meta/task/MergePluginMeta.kt
@@ -53,7 +53,7 @@ public abstract class MergePluginMeta @Inject internal constructor(
             name = meta.name.get(),
             description = meta.description.orNull,
             version = meta.version.get(),
-            apiVersion = meta.apiVersion.orNull,
+            apiVersion = meta.apiVersion.orNull?.takeIf { it.isNotEmpty() },
             website = meta.url.orNull,
             authors = meta.authors.get().takeIf { it.isNotEmpty() }
         )